+++ examples/readme.md
@@ -1,0 +1,16 @@
+# Examples
+
+## ball
+```
+go run github.com/setanarut/ebitencm/examples/ball@latest
+```
+
+## basic
+```
+go run github.com/setanarut/ebitencm/examples/basic@latest
+```
+
+## bench
+```
+go run github.com/setanarut/ebitencm/examples/bench@latest
+```
